Add files via upload
The .model3 file is a model you can import into QGIS that allows you to pick your line to snap to and your pre-generated points that will be snapped to the line. This is used for As Builts of fiber laterals to make points every 50ft with the depth of the lateral and how far it is from the pavement. Building off of the station points model developed earlier - which automatically takes a qchainage layer every 10ft of a line, stylizes them and extracts every 50ft point to a new layer -  the As Built model finds the closest point and creates new vertices to meet the closest point on the line in QGIS. The model changes the attribute table of the final as built layer from the points layer metadata by changing the station point's standardization using the fid column. The layer style is set from a local qgis style file and applied automatically.